requirements.txt
discord==2.2.3
google-api-python-client==2.88.0